Make GPG path customizable
Allow for specifying custom names of the GnuPG executable.  This is
primarily intended to be used when GnuPG executable file name has
a custom suffix (e.g. gpg2, gpg-custom).  The absolute path to GnuPG
executable can be specified this way, but usually altering PATH
environment variable is a cleaner solution.